Growth

Public API

Ecophys.Growth.Organ — Type

Organ(f = [0.53, 0.25, 0.05, 0.05, 0.06, 0.06],
      te = 100.0, tm = 50.0, tb = 10.0, wmax = 0.1)

Data structure to store all parameters for the growth model of a plant organ.

Arguments

f: fraction of dry matter allocated to different carbon pools of newly formed material, [carbohydrates, proteins, lipids, lignin, organic acids, minerals]

te: time at which growth rate extinguishes, [DD]

tm: time at which growth rate is maximum, [DD]

tb: base temperature, [°C]

wmax: maximum weight, [g]

Ecophys.Growth.compute_ASRQ — Function

compute_ASRQ(f::Vector{Float64})

Compute assimilate requirement, CO2 production factor and carbon content according to Penning de Vries et al. 1982. These values can be used to compute growth respiration. Different tissues (species/organ) have different composition. Growth costs are computed given the fraction of dry matter allocated to different carbon pools according to specific composition of the tissue: f = [carbohydrates, proteins, lipids, lignin, organic acids, minerals]

The assimilate requirement (ASRQ, g) is the amount of assimilates required to produce 1 g of dry matter.
The CO2 production factor (CO2PFF, g) is the amount of CO2 produced per g of dry matter.
The carbon fraction (CF, %) is the fraction of dry matter that is carbon.

Citations: Penning de Vries, F. W. T., & van Laar, H. H. (1982). Simulation of growth processes and the model BACROS. In F. W. T. Penning de Vries, & H. H. van Laar (Eds.), Simulation of plant growth and crop production (pp. 114-135). (Simulation monographs). Pudoc. https://edepot.wur.nl/172216

Ecophys.Growth.compute_potential_GR — Function

compute_potential_GR(par::OrganType)

Compute potential growth rate of the leaf according to thermal time (Yin et al 2003) The potential growth rate is computed as a function of thermal time (tt, DD), and the parameters of the growth model: - time at which growth rate extinguishes (te, DD), - time at which growth rate is maximum (tm, DD) - base thermal time at which growth starts (tb, DD) - maximum weight (wmax, g)

Citations: Yin, Xinyou & Goudriaan, Jan & Lantinga, Egbert & Vos, Jan & Spiertz, Huub. (2003). A Flexible Sigmoid Function of Determinate Growth. Annals of botany. 91. 361-71. 10.1093/aob/mcg029.

Private

Private functions, types or constants from Growth. These are not exported, so you need to prefix the function name with Ecophys.Growth to access them. Also bear in mind that these are not part of the public API, so they may change without notice.