Scenes and 3D meshes

Public API

PlantGeomPrimitives.Mesh — Type

Mesh

A struct representing a 3D mesh. Every three vertices represents a triangle.

Fields

vertices: A vector containing the vertices of the mesh.

normals: A vector containing the normals of the mesh.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)];

julia> n = [Vec(0.0, 0.0, 1.0)];

julia> m = Mesh(v, n);

PlantGeomPrimitives.Mesh — Method

Mesh(meshes)

Merge multiple meshes into a single one

Arguments

meshes: Vector of meshes to merge.

Returns

A new Mesh object that is the result of merging all the input meshes.

Example

julia> e = Ellipse(length = 2.0, width = 2.0, n = 10);

julia> r = Rectangle(length = 10.0, width = 0.2);

julia> m = Mesh([e,r]);

PlantGeomPrimitives.Mesh — Method

Mesh(nt, type)

Generate a triangular dense mesh with enough memory allocated to store nt triangles. The behaviour is equivalent to generating an empty mesh but may be computationally more efficient when appending a large number of primitives. If a lower floating precision is required, this may be specified as an optional third argument as in Mesh(10, Float32).

Arguments

nt: The number of triangles to allocate memory for.

type: The floating-point precision type for the mesh data (default is Float64).

Returns

A Mesh object with no vertices or normals.

Example

julia> m = Mesh(1_000);

julia> nvertices(m);

julia> ntriangles(m);

julia> Mesh(1_000, Float32);

PlantGeomPrimitives.Mesh — Method

Mesh(type = Float64)

Generate an empty triangular dense mesh that represents a primitive or 3D scene. By default a Mesh object will only accept coordinates in double floating precision (Float64) but a lower precision can be generated by specifying the corresponding data type as in Mesh(Float32).

Arguments

type: The floating-point precision type for the mesh data (default is Float64).

Returns

A Mesh object with no vertices or normals.

Example

julia> m = Mesh();

julia> nvertices(m);

julia> ntriangles(m);

julia> Mesh(Float32);

PlantGeomPrimitives.Scene — Type

Scene

A scene is a collection of triangular meshes, colors, and materials. It is used for rendering and ray tracing. The scene object is a container for the following fields:

mesh: A triangular mesh (object of type PlantGeomPrimitives.Mesh).

colors: A vector of colors (any type that inherits from ColorTypes.Colorant).

material_ids: A vector of integers that represent the material associated with each

triangle in the mesh.

materials: A vector of materials (object of types that inherit from PlantGeomPrimitives.Material).

Several constructors are available to create a scene from scratch or to extend an existing scene.

PlantGeomPrimitives.Scene — Method

Scene(scenes)

Merge multiple scenes into a single one.

Arguments

scenes: A vector of scene objects (object of type Scene).

Example

```jldoctest julia> t1 = Triangle(length = 1.0, width = 1.0);

julia> t2 = Rectangle(length = 5.0, width = 0.5);

julia> s1 = Scene(mesh = t1);

julia> s2 = Scene(mesh = t2);

julia> s = Scene([s1, s2]);

PlantGeomPrimitives.Scene — Method

Scene(; mesh = Mesh(Float64), colors = nothing, materials = nothing)

Create a scene from a triangular mesh and associated colors and materials. This function should only be used when creating a scene from scratch. To extend an existing scene, use the function add! instead. Also, scenes can be created from PlantGraphs.Graph objects directly.

Arguments

mesh: A triangular mesh (object of type PlantGeomPrimitives.Mesh).

colors: A vector of colors (any type that inherits from ColorTypes.Colorant). There

should be one color per triangle in the mesh or a single color (such that all triangles get the same color). This is an optional argument, but if no colors are provided, it the resulting scene cannot be visualized (the function rennder() will throw an error).

materials: A vector of materials (object of type PlantRayTracer.Material). There should

be one material per triangle in the mesh or a single material (such that all triangles get the same material). This is an optional argument, but if no materials are provided, the resulting scene cannot be used for ray tracing.

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

PlantGeomPrimitives.Vec — Type

Vec(x, y, z)

3D vector or point with coordinates x, y and z.

julia> v = Vec(0.0, 0.0, 0.0);

julia> v = Vec(0f0, 0f0, 0f0);

PlantGeomPrimitives.BBox — Method

BBox(pmin::Vec, pmax::Vec)

Build an axis-aligned bounding box given the vector of minimum (pmin) and maximum (pmax) coordinates.

Arguments

pmin: The minimum coordinates of the bounding box.

pmax: The maximum coordinates of the bounding box.

Examples

julia> p0 = Vec(0.0, 0.0, 0.0);

julia> p1 = Vec(1.0, 1.0, 1.0);

julia> box = BBox(p0, p1);

PlantGeomPrimitives.BBox — Method

BBox(m::Mesh)

Build a tight axis-aligned bounding box around a Mesh object.

Arguments

m: The mesh to build the bounding box around.

Examples

julia> m = Rectangle();

julia> box = BBox(m);

PlantGeomPrimitives.Ellipse — Method

Ellipse(;length = 1.0, width = 1.0, n = 20)

Create a triangular mesh approximating an ellipse with dimensions given by length and width, discretized into n triangles (must be even) and standard location and orientation.

Arguments

length = 1.0: The length of the ellipse.

width = 1.0: The width of the ellipse.

n = 20: The number of triangles to be used in the mesh.

Examples

julia> Ellipse(;length = 1.0, width = 1.0, n = 20);

PlantGeomPrimitives.HollowCone — Method

HollowCone(;length = 1.0, width = 1.0, height = 1.0, n = 20)

Create a hollow cone with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Arguments

length = 1.0: The length of the cone (distance between base and apex).

width = 1.0: The width of the base of the cone.

height = 1.0: The height of the base of the cone.

n = 20: The number of triangles to be used in the mesh.

Examples

julia> HollowCone(;length = 1.0, width = 1.0, height = 1.0, n = 20);

PlantGeomPrimitives.HollowCube — Method

HollowCube(;length = 1.0, width = 1.0, height = 1.0)

Create a hollow cube (a prism) with dimensions given by length, width and height, standard location and orientation.

Arguments

length = 1.0: The length of the cube.

width = 1.0: The width of the base of the cube.

height = 1.0: The height of the base of the cube.

Examples

julia> HollowCube(;length = 1.0, width = 1.0, height = 1.0);

PlantGeomPrimitives.HollowCylinder — Method

HollowCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 40)

Create a hollow cylinder with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Arguments

length = 1.0: The length of the cylinder (distance between bases).

width = 1.0: The width of the base of the cylinder.

height = 1.0: The height of the base of the cylinder.

n = 40: The number of triangles to discretize the cylinder into.

Examples

julia> HollowCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.HollowFrustum — Method

HollowFrustum(;length = 1.0, width = 1.0, height = 1.0, ratio = 1.0, n = 40)

Create a hollow frustum with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Arguments

length = 1.0: The length of the frustum (distance between bases).

width = 1.0: The width of the base of the frustum.

height = 1.0: The height of the base of the frustum.

ratio = 1.0: The ratio between the top and bottom base radii.

n = 40: The number of triangles to discretize the frustum into.

Examples

julia> HollowFrustum(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.O — Method

O()

Returns the origin of the 3D coordinate system as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in O(Float32).

julia>  O();

julia>  O(Float32);

PlantGeomPrimitives.Rectangle — Method

Rectangle(;length = 1.0, width = 1.0)

Create a rectangle with dimensions given by length and width, standard location and orientation.

Arguments

length = 1.0: The length of the rectangle.

width = 1.0: The width of the rectangle.

Examples

julia> Rectangle(;length = 1.0, width = 1.0);

PlantGeomPrimitives.SolidCone — Method

SolidCone(;length = 1.0, width = 1.0, height = 1.0, n = 40)

Create a solid cone with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Arguments

length = 1.0: The length of the cone (distance between base and apex).

width = 1.0: The width of the base of the cone.

height = 1.0: The height of the base of the cone.

n = 40: The number of triangles to be used in the mesh.

Examples

julia> SolidCone(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.SolidCube — Method

SolidCube(;length = 1.0, width = 1.0, height = 1.0)

Create a solid cube with dimensions given by length, width and height, standard location and orientation.

Arguments

length = 1.0: The length of the cube.

width = 1.0: The width of the base of the cube.

height = 1.0: The height of the base of the cube.

Examples

julia> SolidCube(;length = 1.0, width = 1.0, height = 1.0);

PlantGeomPrimitives.SolidCylinder — Method

SolidCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 80)

Create a solid cylinder with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Arguments

length = 1.0: The length of the cylinder (distance between bases).

width = 1.0: The width of the base of the cylinder.

height = 1.0: The height of the base of the cylinder.

n = 80: The number of triangles to discretize the cylinder into.

Examples

julia> SolidCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 80);

PlantGeomPrimitives.SolidFrustum — Method

SolidFrustum(;length = 1.0, width = 1.0, height = 1.0, ratio = 1.0, n = 40)

Create a solid frustum with dimensions given by length, width and height, discretized into n triangles and standard location and orientation.

Arguments

length = 1.0: The length of the frustum (distance between bases).

width = 1.0: The width of the base of the frustum.

height = 1.0: The height of the base of the frustum.

ratio = 1.0: The ratio between the top and bottom base radii.

n = 40: The number of triangles to discretize the frustum into.

Examples

julia> SolidFrustum(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.Trapezoid — Method

Trapezoid(;length = 1.0, width = 1.0, ratio = 1.0)

Create a trapezoid with dimensions given by length and the larger width and the ratio between the smaller and larger widths. The trapezoid is generted at the standard location and orientation.

Arguments

length = 1.0: The length of the trapezoid.

width = 1.0: The larger width of the trapezoid (the lower base of the trapezoid).

ratio = 1.0: The ratio between the smaller and larger widths.

Examples

julia> Trapezoid(;length = 1.0, width = 1.0, ratio = 1.0);

PlantGeomPrimitives.Triangle — Method

Triangle(;length = 1.0, width = 1.0)

Create a triangle with dimensions given by length and width, standard location and orientation.

Arguments

length = 1.0: The length of the triangle.

width = 1.0: The width of the triangle.

Examples

julia> Triangle(;length = 1.0, width = 1.0);

PlantGeomPrimitives.X — Method

X(s)

Returns scaled vector in the direction of the X axis with length s as a Vec object using the same floating point precision as s.

julia>  X(1.0);

julia>  X(1f0) ;

PlantGeomPrimitives.X — Method

X()

Returns an unit vector in the direction of the X axis as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in X(Float32).

julia>  X();

julia>  X(Float32);

PlantGeomPrimitives.Y — Method

Y(s)

Returns scaled vector in the direction of the Y axis with length s as a Vec object using the same floating point precision as s.

julia>  Y(1.0);

julia>  Y(1f0);

PlantGeomPrimitives.Y — Method

Y()

Returns an unit vector in the direction of the Y axis as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in Y(Float32).

julia>  Y();

julia>  Y(Float32);

PlantGeomPrimitives.Z — Method

Z(s)

Returns scaled vector in the direction of the Z axis with length s as a Vec object using the same floating point precision as s.

julia>  Z(1.0);

julia>  Z(1f0);

PlantGeomPrimitives.Z — Method

Z()

Returns an unit vector in the direction of the Z axis as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in Z(Float32).

julia>  Z();

julia>  Z(Float32);

PlantGeomPrimitives.add! — Method

add!(scene; mesh, colors = nothing, materials = nothing)

Manually add a mesh to an existing scene with optional colors and materials. Make sure to be consistent with the optional arguments. That is, if the scene was created with colors, then you should provide colors for the new mesh as well (the same applies to materials). Otherwise, the scene will not be usable for rendering or ray tracing.

Arguments

scene: A scene object.

mesh: A triangular mesh.

colors: A vector of colors. See documentation of Scene for more information.

materials: A vector of materials. See documentation of Scene for more information.

Example

```jldoctest julia> t1 = Triangle(length = 1.0, width = 1.0);

julia> t2 = Rectangle(length = 5.0, width = 0.5);

julia> s1 = Scene(mesh = t1);

julia> add!(s1, mesh = t2);

julia> ntriangles(s1);

PlantGeomPrimitives.area — Method

area(mesh::Mesh)

Total surface area of a mesh (as the sum of areas of individual triangles).

Arguments

mesh: Mesh which area is to be calculated.

Returns

The total surface area of the mesh as a number.

Example

julia> r = Rectangle(length = 10.0, width = 0.2);

julia> area(r);

julia> r = Rectangle(length = 10f0, width = 0.2f0);

julia> area(r);

PlantGeomPrimitives.areas — Method

areas(m::Mesh)

A vector with the areas of the different triangles that form a mesh.

Arguments

mesh: Mesh which areas are to be calculated.

Returns

A vector with the areas of the different triangles that form the mesh.

Example

julia> r = Rectangle(length = 10.0, width = 0.2);

julia> areas(r);

julia> r = Rectangle(length = 10f0, width = 0.2f0);

julia> areas(r);

PlantGeomPrimitives.colors — Method

colors(scene)

Extract the vector of colors stored inside a scene (used for rendering). Note that a scene will assign one color per vertex (even though we create scenes with one color per triangle or per mesh).

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> import ColorTypes: RGB

julia> s = Scene(mesh = t, colors = RGB(1.0, 0.0, 0.0));

julia> colors(s);

julia> s = Scene(mesh = t);

julia> colors(s);

PlantGeomPrimitives.get_triangle — Method

get_triangle(m::Mesh, i)

Retrieve the vertices for the i-th triangle in a mesh.

Arguments

mesh: The mesh from which to retrieve the triangle.

i: The index of the triangle to retrieve.

Returns

A vector containing the three vertices defining the i-th triangle.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0),
            Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(0.0, 0.0, 1.0)];

julia> n = [Vec(0.0, 0.0, 1.0),Vec(0.0, 0.0, -1.0)];

julia> m = Mesh(v, n);

julia> get_triangle(m, 2);

PlantGeomPrimitives.load_mesh — Method

load_mesh(filename, type = Float64)

Import a mesh from a file given by filename. Supported formats include stl, ply, obj and msh. By default, this will generate a Mesh object that uses double floating-point precision. However, a lower precision can be specified by passing the relevant data type as in load_mesh(filename, Float32).

Arguments

filename: The path to the file containing the mesh.

type: The floating-point precision type for the mesh data (default is Float64).

Example

julia> mesh = load_mesh("path/to/mesh.obj")
Mesh{StaticArraysCore.SVector{3, Float64}}

julia> mesh = load_mesh("path/to/mesh.obj", Float32)
Mesh{StaticArraysCore.SVector{3, Float32}}

PlantGeomPrimitives.materials — Method

materials(scene)

Extract the vector of material objects stored inside a scene (used for ray tracing). Depending on how the scene was created, there may be one material per triangle or one material per mesh inside the scene.

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> import PlantRayTracer: Sensor;

julia> s = Scene(mesh = t, materials = Sensor(1));

julia> materials(s);

julia> s = Scene(mesh = t);

julia> materials(s);

PlantGeomPrimitives.mesh — Method

mesh(scene)

Extract the triangular mesh stored inside a scene (used for ray tracing & rendering).

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

julia> mesh(s);

PlantGeomPrimitives.normals — Method

normals(mesh::Mesh)

Retrieve the normals of a mesh.

Arguments

mesh: The mesh from which to retrieve the normals.

Returns

A vector containing the normals of the mesh.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)];

julia> n = [Vec(0.0, 0.0, 1.0)];

julia> m = Mesh(v, n);

julia> normals(m);

PlantGeomPrimitives.normals — Method

normals(scene)

Extract the normals of the triangular mesh stored inside a scene.

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

julia> normals(s);

PlantGeomPrimitives.ntriangles — Method

ntriangles(mesh)

Extract the number of triangles in a mesh.

Arguments

mesh: The mesh from which to extract the number of triangles.

Returns

The number of triangles in the mesh as an integer.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)];

julia> n = [Vec(0.0, 0.0, 1.0)];

julia> m = Mesh(v, n);

julia> ntriangles(m);

PlantGeomPrimitives.ntriangles — Method

ntriangles(scene)

Calculate the number of vertices

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

julia> ntriangles(s);

PlantGeomPrimitives.nvertices — Method

nvertices(mesh)

The number of vertices in a mesh.

Arguments

mesh: The mesh from which to retrieve the number of vertices.

Returns

The number of vertices in the mesh as an integer.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)];

julia> n = [Vec(0.0, 0.0, 1.0)];

julia> m = Mesh(v, n);

julia> nvertices(m);

PlantGeomPrimitives.nvertices — Method

nvertices(scene)

Calculate the number of vertices

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

julia> nvertices(s);

PlantGeomPrimitives.rotate! — Method

rotate!(m::Mesh; x::Vec, y::Vec, z::Vec)

Rotate a mesh m to a new coordinate system given by x, y and z

PlantGeomPrimitives.rotatex! — Method

rotatex!(m::Mesh, θ)

Rotate a mesh m around the x axis by angle θ.

Arguments

m: The mesh to be scaled.

θ: Angle of rotation in radians.

Examples

julia> m = Rectangle();

julia> θ = pi/2;

julia> rotatex!(m, θ)

PlantGeomPrimitives.rotatey! — Method

rotatey!(m::Mesh, θ)

Rotate a mesh m around the y axis by angle θ.

Arguments

m: The mesh to be scaled.

θ: Angle of rotation in radians.

Examples

julia> m = Rectangle();

julia> θ = pi/2;

julia> rotatey!(m, θ);

PlantGeomPrimitives.rotatez! — Method

rotatez!(m::Mesh, θ)

Rotate a mesh m around the z axis by angle θ.

Arguments

m: The mesh to be scaled.

θ: Angle of rotation in radians.

Examples

julia> m = Rectangle();

julia> θ = pi/2;

julia> rotatez!(m, θ);

PlantGeomPrimitives.save_mesh — Method

save_mesh(mesh; fileformat = :STL_BINARY, filename)

Save a mesh into an external file using a variety of formats.

Arguments

mesh: Object of type Mesh.

fileformat: Format to store the mesh as symbol.

filename: Name of the file in which to store the mesh as string.

Details

The fileformat should take one of the following arguments: :STL_BINARY, :STL_ASCII, :PLY_BINARY, :PLY_ASCII or :OBJ. Note that these names should be passed as symnols.

Example

v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)]
n = [Vec(0.0, 0.0, 1.0)]
mesh = Mesh(v, n)
save_mesh(mesh, fileformat = :STL_BINARY, filename = "path/to/mesh.bstl")

PlantGeomPrimitives.scale! — Method

scale!(m::Mesh, vec::Vec)

Scale a mesh m along the three axes provided by vec.

Arguments

m: The mesh to be scaled.

vec: A vector containing the scaling factors for the x, y, and z axes.

Examples

julia> m = Rectangle();

julia> scaling_vector = Vec(2.0, 1.5, 3.0);

julia> scale!(m, scaling_vector);

PlantGeomPrimitives.translate! — Method

translate!(m::Mesh, v::Vec)

Translate the mesh m by vector v.

Arguments

m: The mesh to be translated.

v: The vector by which the mesh is to be translated.

Examples

julia> m = Rectangle();

julia> v = Vec(2.0, 1.5, 3.0);

julia> translate!(m, v);

PlantGeomPrimitives.vertices — Method

vertices(mesh::Mesh)

Retrieve the vertices of a mesh.

Arguments

mesh: The mesh from which to retrieve the vertices.

Returns

A vector containing the vertices of the mesh.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)];

julia> n = [Vec(0.0, 0.0, 1.0)];

julia> m = Mesh(v, n);

julia> vertices(m);

PlantGeomPrimitives.vertices — Method

vertices(scene)

Extract the vertices of the triangular mesh stored inside a scene.

Arguments

scene: A scene object (object of type Scene).

Example

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

julia> vertices(s);

Private

Private functions, types or constants from PlantGeomPrimitives. These are not exported, so you need to prefix the function name with PlantGeomPrimitives. to access them. Also bear in mind that these are not part of the public API, so they may change without notice.

Base.eltype — Method

eltype(mesh::Mesh)

Extract the the type used to represent coordinates in a mesh (e.g., Float64).

Fields

mesh: The mesh from which to extract the element type.

Example

julia> v = [Vec(0.0, 0.0, 0.0), Vec(0.0, 1.0, 0.0), Vec(1.0, 0.0, 0.0)];

julia> n = [Vec(0.0, 0.0, 1.0)];

julia> m = Mesh(v, n);

julia> eltype(m);