Turtle Geometry

API documentation for PlantGeomTurtle.jl.

Turtle geometry

PlantGeomTurtle.Turtle — Method

Turtle(Float64, message)

Create a meshing turtle that can convert a Graph into a 3D mesh using turtle operators, geometry primitives and methods of feed!(). By default, the meshing turtle will generate geometry primitives with double floating precision (Float64) but it is possible to generate a version with lower precision as in Turtle(Float32). The argument message is any user-defined object.

PlantGeomTurtle.head — Method

head(turtle)

Extract the direction vector (a Vec object) of the head of the turtle.

PlantGeomTurtle.up — Method

up(turtle)

Extract the direction vector (a Vec object) of the back of the turtle.

PlantGeomTurtle.arm — Method

arm(turtle)

Extract the direction vector (a Vec object) of the arm of the turtle.

PlantGeomTurtle.pos — Method

pos(turtle)

Extract the current position of the turtle (a Vec object).

PlantGeomTurtle.geoms — Method

geoms(turtle)

Extract the 3D mesh generated by the turtle (a Mesh object).

PlantGeomPrimitives.colors — Method

colors(turtle)

Extract the color objects associated to each geometry primitive that was fed to the turtle.

PlantGeomPrimitives.faces — Method

geoms(turtle)

Extract the faces of the 3D mesh generated by the turtle.

PlantGeomPrimitives.materials — Method

materials(turtle)

Extract the material objects associated to each geometry primitive that was fed to the turtle.

PlantGeomTurtle.feed! — Function

feed!(turtle::Turtle; mesh::Mesh, color::Colorant = nothing, mat::Material = nothing)

General purpose method to feed a mesh to a turtle together with color and material. Note that all primitives provided by VPL are implemented as meshes, but this is a generic method for meshes that are constructed directly by the user or imported from external software.

feed!(turtle::Turtle, node::Node, vars = nothing)

Default method for feed!() that does not do anything. This allows the user to include nodes in a graph without an associated geometry.

feed!(turtle::Turtle, g::Graph)

Process a Graph object with a turtle and generate the corresponding 3D mesh from executing the different feed!() methods associated to the nodes in the graph.

feed!(turtle::Turtle, collection::AbstractArray)
feed!(turtle::Turtle, collection::Tuple)

Feed a turtle an array or tuple of objects (collection) with existing feed!() methods.

PlantGeomPrimitives.Scene — Method

Scene(graph, Float64)

Create a 3D scene from a Graph object (g). By default, double floating precision will be used (Float64) but it is possible to generate a version with a different precision by specifying the corresponding type as in Scene(g, Float32). The Scene object contains a mesh of triangles as well as colors and materials associated to each primitive.

PlantGeomPrimitives.Scene — Method

Scene(graphs, Float64; parallel = false, message = nothing)

Create a 3D scene for rendering from an array of Graph objects (graphs). The graphs may be processed serially (default) or in parallel using multithreading (parallel = true). By default, double floating precision will be used (Float64) but it is possible to generate a version with a different precision by specifying the corresponding type as in Scene(graphs, Float32).

PlantGeomTurtle.T — Type

T(to::Vec)

Node that translates a turtle to the new position to (a Vec object).

PlantGeomTurtle.t! — Method

t!(turtle; to = O())

Translate a turtle to the new position to (a Vec object).

Examples

julia> turtle = Turtle();

julia> using PlantGeomPrimitives

julia> t!(turtle, to = Y(1.0));

PlantGeomTurtle.OR — Type

OR(head::Vec, up::Vec, arm::Vec)

Node that orients a turtle to a new direction by re-defining the local reference system.

PlantGeomTurtle.or! — Method

or!(turtle; head = Z(), up = X(), arm = Y())

Orient a turtle to a new direction by re-defining the local reference system. The arguments head, up and arm should be of type Vec.

Examples

julia> turtle = Turtle();

julia> using PlantGeomPrimitives

julia> or!(turtle, head = Y(), up = Z(), arm = X());

PlantGeomTurtle.SET — Type

SET(to, head, up, arm)

Node that sets the position and orientation of a turtle.

PlantGeomTurtle.set! — Method

set!(turtle; to = O(), head = Z(), up = X(), arm = Y())

Set position and orientation of a turtle. The arguments to, head, up and arm should be of type Vec and be passed as keyword arguments.

Examples

julia> turtle = Turtle();

julia> using PlantGeomPrimitives

julia> set!(turtle, to = O(), head = Y(), up = Z(), arm = X());

PlantGeomTurtle.RU — Type

RU(angle)

Node that rotates a turtle around up axis. Angle must be in hexadecimal degrees and the rotation is clockwise.

PlantGeomTurtle.ru! — Method

ru!(turtle, angle)

Rotates a turtle around up axis. Angle must be in hexadecimal degrees and the rotation is clockwise.

Examples

julia> turtle = Turtle();

julia> ru!(turtle, 45.0);

PlantGeomTurtle.RA — Type

RA(angle)

Node that rotates a turtle around arm axis. Angle must be in hexadecimal degrees and the rotation is clockwise.

PlantGeomTurtle.ra! — Method

ra!(turtle, angle)

Rotates a turtle around arm axis. Angle must be in hexadecimal degrees and the rotation is clockwise.

Examples

julia> turtle = Turtle();

julia> ra!(turtle, 45.0);

PlantGeomTurtle.RH — Type

RH(angle)

Node that rotates a turtle around head axis. Angle must be in hexadecimal degrees and the rotation is clockwise.

PlantGeomTurtle.rh! — Method

rh!(turtle, angle)

Rotate turtle around head axis. Angle must be in hexadecimal degrees and the rotation is clockwise.

Examples

julia> turtle = Turtle();

julia> rh!(turtle, 45.0);

PlantGeomTurtle.F — Type

F(dist)

Moves a turtle forward a given distance.

PlantGeomTurtle.f! — Method

f!(turtle, dist)

Move turtle forward a given distance.

Examples

julia> turtle = Turtle();

julia> f!(turtle, 2.0);

PlantGeomTurtle.RV — Type

RV(strength)

Rotates the turtle towards the Z axis. See documentation for rv! for details.

PlantGeomTurtle.rv! — Method

rv!(turtle, strength)

Rotates the turtle towards the Z axis. The angle of rotation is proportional to the cosine of the zenith angle of the turtle (i.e., angle between its head and the vertical axis) with the absolute value of strength being the proportion between the two. strength should vary between -1 and 1. If strength is negative, the turtle rotates downwards (i.e., towards negative values of Z axis), otherwise upwards.

Examples

julia> turtle = Turtle();

julia> ra!(turtle, 45.0);

julia> rv!(turtle, 0.5);

Geometry primitives

Triangle

PlantGeomPrimitives.Triangle! — Method

Triangle!(turtle; length = 1.0, width = 1.0, move = false,
          material = nothing, color = nothing)

Generate a triangle in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the triangle to.

length: Length of the triangle.

width: Width of the triangle.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A triangle mesh will be generated representing the triangle. The triangle will be generated in front of the turtle, on the plane defined by the arm and head axes of the turtle. The argument length refers to the axis of the triangle aligned with the head axis of the turtle, whereas width refers to the orthogonal axis.

When move = true, the turtle will be moved forward by a distance equal to length.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> Triangle!(turtle; length = 2.0, width = 1.0);

Rectangle

PlantGeomPrimitives.Rectangle! — Method

Rectangle!(turtle; length = 1.0, width = 1.0, move = false,
           material = nothing, color = nothing)

Generate a rectangle in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the rectangle to.

length: Length of the rectangle.

width: Width of the rectangle.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A triangle mesh will be generated representing the rectangle. The rectangle will be generated in front of the turtle, on the plane defined by the arm and head axes of the turtle. The argument length refers to the axis of the rectangle aligned with the head axis of the turtle, whereas width refers to the orthogonal axis.

When move = true, the turtle will be moved forward by a distance equal to length.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> Rectangle!(turtle; length = 1.0, width = 0.5);

Trapezoid

PlantGeomPrimitives.Trapezoid! — Method

Trapezoid!(turtle; length = 1.0, width = 1.0, ratio = 1.0, move = false,
material = nothing, color = nothing)

Generate a trapezoid in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the trapezoid to.

length: Length of the trapezoid.

width: Width of the base of the trapezoid.

ratio: Ratio between the width of the top and base of the trapezoid.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A triangle mesh will be generated representing the trapezoid. The trapezoid will be generated in front of the turtle, on the plane defined by the arm and head axes of the turtle. The argument length refers to the axis of the trapezoid aligned with the head axis of the turtle, whereas width refers to the orthogonal axis.

When move = true, the turtle will be moved forward by a distance equal to length.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> Trapezoid!(turtle; length = 1.0, width = 1.0, ratio = 0.5);

Ellipse

PlantGeomPrimitives.Ellipse! — Method

Ellipse!(turtle; length = 1.0, width = 1.0, n = 20, move = false,
         material = nothing, color = nothing)

Generate an ellipse in front of a turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the ellipse to.

length: Length of the ellipse.

width: Width of the ellipse.

n: Number of triangles of the mesh approximating the ellipse (an integer).

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A triangle mesh will be generated with n triangles that approximates an ellipse. The ellipse will be generated in front of the turtle, on the plane defined by the arm and head axes of the turtle. The argument length refers to the axis of the ellipse aligned with the head axis of the turtle, whereas width refers to the orthogonal axis.

When move = true, the turtle will be moved forward by a distance equal to length.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> Ellipse!(turtle; length = 1.0, width = 0.5, n = 40);

Hollow cylinder

PlantGeomPrimitives.HollowCylinder! — Method

HollowCylinder!(turtle; length = 1.0, width = 1.0, height = 1.0, n = 40, move = false,
material = nothing, color = nothing)

Generate a hollow cylinder in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the hollow cylinder to.

length: Length of the ellipse at the base of the hollow cylinder.

width: Width of the ellipse at the base of the hollow cylinder.

height: Height of the hollow cylinder.

n: Number of triangles in the mesh (must be even).

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated with n triangles that approximate the hollow cylinder. The cylinder will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> HollowCylinder!(turtle; length = 1.0, width = 1.0, height = 2.0, n = 40);

PlantGeomPrimitives.SolidCylinder! — Method

SolidCylinder!(turtle; length = 1.0, width = 1.0, height = 1.0, n = 80, move = false,
material = nothing, color = nothing)

Generate a solid cylinder in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the solid cylinder to.

length: Length of the ellipse at the base of the solid cylinder.

width: Width of the ellipse at the base of the solid cylinder.

height: Height of the solid cylinder.

n: Number of triangles in the mesh (must be even).

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated with n triangles that approximate the solid cylinder. The cylinder will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> SolidCylinder!(turtle; length = 1.0, width = 1.0, height = 2.0, n = 80);

Hollow cone

PlantGeomPrimitives.HollowCone! — Method

HollowCone!(turtle; length = 1.0, width = 1.0, height = 1.0, n = 20, move = false,
material = nothing, color = nothing)

Generate a hollow cone in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the hollow cone to.

length: Length of the ellipse at the base of the hollow cone.

width: Width of the ellipse at the base of the hollow cone.

height: Height of the hollow cone.

n: Number of triangles in the mesh.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated with n triangles that approximate the hollow cone. The cone will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> Trapezoid!(turtle; length = 1.0, width = 1.0, ratio = 0.5);

PlantGeomPrimitives.SolidCone! — Method

SolidCone!(turtle; length = 1.0, width = 1.0, height = 1.0, n = 40, move = false,
material = nothing, color = nothing)

Generate a solid frustum in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the solid cone to.

length: Length of the ellipse at the base of the solid cone.

width: Width of the ellipse at the base of the solid cone.

height: Height of the solid cone.

n: Number of triangles in the mesh (must be even).

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated with n triangles that approximate the solid cone. The cone will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> SolidCone!(turtle; length = 1.0, width = 1.0, height = 2.0, n = 40);

Cube

PlantGeomPrimitives.SolidCube! — Method

SolidCube!(turtle; length = 1.0, width = 1.0, height = 1.0, move = false,
material = nothing, color = nothing)

Generate a solid cube in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the solid cube to.

length: Length of the rectangle at the base of the solid cube.

width: Width of the rectangle at the base of the solid cube.

height: Height of the solid cube.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated of a solid cube. The cube will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> SolidCube!(turtle; length = 1.0, width = 1.0, height = 2.0);

PlantGeomPrimitives.HollowCube! — Method

HollowCube!(turtle; length = 1.0, width = 1.0, height = 1.0, move = false,
material = nothing, color = nothing)

Generate a hollow cube in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the hollow cube to.

length: Length of the rectangle at the base of the hollow cube.

width: Width of the rectangle at the base of the hollow cube.

height: Height of the hollow cube.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated of a hollow cube. The cube will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> HollowCube!(turtle; length = 1.0, width = 1.0, height = 2.0);

Solid frustum

PlantGeomPrimitives.SolidFrustum! — Method

SolidFrustum!(turtle; length = 1.0, width = 1.0, height = 1.0, n = 80, move = false,
material = nothing, color = nothing)

Generate a solid frustum in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the solid frustum to.

length: Length of the ellipse at the base of the solid frustum.

width: Width of the ellipse at the base of the solid frustum.

height: Height of the solid frustum.

n: Number of triangles in the mesh (must be even).

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated with n triangles that approximate the solid frustum. The frustum will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> SolidFrustum!(turtle; length = 1.0, width = 1.0, height = 2.0, n = 80);

PlantGeomPrimitives.HollowFrustum! — Method

HollowFrustum!(turtle; length = 1.0, width = 1.0, height = 1.0, n = 40, move = false,
material = nothing, color = nothing)

Generate a hollow frustum in front of the turtle and feed it to a turtle.

Arguments

turtle: The turtle that we feed the hollow frustum to.

length: Length of the ellipse at the base of the hollow frustum.

width: Width of the ellipse at the base of the hollow frustum.

height: Height of the hollow frustum.

n: Number of triangles in the mesh (must be even).

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A mesh will be generated with n triangles that approximate the hollow frustum. The frustum will be generated in front of the turtle, with the base on the plane defined by the arm and up axes of the turtle, centered at the head axis. The length argument refers to the up axis, whereas width refers to the arm axis and height is associated to the head axis.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> turtle = Turtle();

julia> HollowFrustum!(turtle; length = 1.0, width = 1.0, height = 2.0, n = 40);

Generic mesh

PlantGeomTurtle.Mesh! — Method

Mesh!(turtle, m::Mesh; scale = Vec(1.0, 1.0, 1.0), move = false,
material = nothing, color = nothing)

Feed a pre-existing mesh to a turtle after scaling.

Arguments

turtle: The turtle that we feed the mesh to.

m: The pre-existing unscaled mesh in standard position and orientation.

scale: Vector with scaling factors for the x, y and z axes.

move: Whether to move the turtle forward or not (true or false).

material: The material object for the ray tracer (optional).

color: The color of the ellipse for rendering (optional).

Details

A pre-existing mesh will be scaled (acccording to scale), rotate so that it is oriented in the same direction as the turtle and translated so that the mesh is generated in front of the turtle. A deep copy of the original mesh is made prior to any transformation.

When move = true, the turtle will be moved forward by a distance equal to height.

The material object must inherit from Material (see ray tracing documentation for detail) and the color can be any type that inherits from Colorant (from ColorTypes.jl).

Return

Returns nothing but modifies the turtle as a side effect.

Examples

julia> import PlantGeomPrimitives as PG

julia> e = PG.Ellipse();

julia> turtle = Turtle();

julia> Mesh!(turtle, e, scale = PG.Vec(2.0, 2.0, 2.0));