Scenes and 3D meshes

Public API

PlantGeomPrimitives.Mesh — Method

Mesh(nt, nv = nt*3)

Generate a triangular dense mesh with enough memory allocated to store nt triangles and nv vertices. The behaviour is equivalent to generating an empty mesh but may be computationally more efficient when appending a large number of primitives. If a lower floating precision is required, this may be specified as an optional third argument as in Mesh(10, 30, Float32).

PlantGeomPrimitives.Mesh — Method

Mesh()

Generate an empty triangular dense mesh that represents a primitive or 3D scene. By default a Mesh object will only accept coordinates in double floating precision (Float64) but a lower precision can be generated by specifying the corresponding data type as in Mesh(Float32).

PlantGeomPrimitives.Scene — Method

Scene(scenes)

Merge multiple Scene objects into one.

PlantGeomPrimitives.Scene — Method

Scene(; mesh = Mesh(Float64), colors = Colorant[], material_ids = Int[], materials = Material[])

Create a Scene object from a triangular mesh (mesh), a vector of colors (colors, any type that inherits from Colorant from the ColorTypes package), a vector of material IDs (material_ids that link indivudal triangles to material objects) and a vector of materials (materials, any object that inherits from Material). See packages PlantViz and PlantRayTracer for more details on materials and colors.

julia> t = Triangle(length = 2.0, width = 2.0);

julia> s = Scene(mesh = t);

PlantGeomPrimitives.Vec — Type

Vec(x, y, z)

3D vector or point with coordinates x, y and z.

julia> v = Vec(0.0, 0.0, 0.0);

julia> v = Vec(0f0, 0f0, 0f0);

PlantGeomPrimitives.BBox — Method

BBox(pmin::Vec, pmax::Vec)

Build an axis-aligned bounding box given the vector of minimum (pmin) and maximum (pmax) coordinates.

Examples

julia> p0 = Vec(0.0, 0.0, 0.0);

julia> p1 = Vec(1.0, 1.0, 1.0);

julia> box = BBox(p0, p1);

PlantGeomPrimitives.BBox — Method

BBox(m::Mesh)

Build a tight axis-aligned bounding box around a Mesh object.

PlantGeomPrimitives.Ellipse — Method

Ellipse(;length = 1.0, width = 1.0, n = 20)

Create an ellipse with dimensions given by length and width, discretized into n triangles (must be even) and standard location and orientation.

Examples

julia> Ellipse(;length = 1.0, width = 1.0, n = 20);

PlantGeomPrimitives.HollowCone — Method

HollowCone(;length = 1.0, width = 1.0, height = 1.0, n = 20)

Create a hollow cone with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Examples

julia> HollowCone(;length = 1.0, width = 1.0, height = 1.0, n = 20);

PlantGeomPrimitives.HollowCube — Method

HollowCube(;length = 1.0, width = 1.0, height = 1.0)

Create a hollow cube with dimensions given by length, width and height, standard location and orientation.

Examples

julia> HollowCube(;length = 1.0, width = 1.0, height = 1.0);

PlantGeomPrimitives.HollowCylinder — Method

HollowCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 40)

Create a hollow cylinder with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Examples

julia> HollowCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.HollowFrustum — Method

HollowFrustum(;length = 1.0, width = 1.0, height = 1.0, n = 40)

Create a hollow frustum with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Examples

julia> HollowFrustum(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.O — Method

O()

Returns the origin of the 3D coordinate system as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in O(Float32).

julia>  O();

julia>  O(Float32);

PlantGeomPrimitives.Rectangle — Method

Rectangle(;length = 1.0, width = 1.0)

Create a rectangle with dimensions given by length and width, standard location and orientation.

Examples

julia> Rectangle(;length = 1.0, width = 1.0);

PlantGeomPrimitives.SolidCone — Method

SolidCone(;length = 1.0, width = 1.0, height = 1.0, n = 40)

Create a solid cone with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Examples

julia> SolidCone(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.SolidCube — Method

SolidCube(;length = 1.0, width = 1.0, height = 1.0)

Create a solid cube with dimensions given by length, width and height, standard location and orientation.

Examples

julia> SolidCube(;length = 1.0, width = 1.0, height = 1.0);

PlantGeomPrimitives.SolidCylinder — Method

SolidCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 80)

Create a solid cylinder with dimensions given by length, width and height, discretized into n triangles (must be even) and standard location and orientation.

Examples

julia> SolidCylinder(;length = 1.0, width = 1.0, height = 1.0, n = 80);

PlantGeomPrimitives.SolidFrustum — Method

SolidFrustum(;length = 1.0, width = 1.0, height = 1.0, n = 40)

Create a solid frustum with dimensions given by length, width and height, discretized into n triangles and standard location and orientation.

Examples

julia> SolidFrustum(;length = 1.0, width = 1.0, height = 1.0, n = 40);

PlantGeomPrimitives.Trapezoid — Method

Trapezoid(;length = 1.0, width = 1.0, ratio = 1.0)

Create a trapezoid with dimensions given by length and the larger width and the ratio between the smaller and larger widths. The trapezoid is generted at the standard location and orientation.

Examples

julia> Trapezoid(;length = 1.0, width = 1.0, ratio = 1.0);

PlantGeomPrimitives.Triangle — Method

Triangle(;length = 1.0, width = 1.0)

Create a triangle with dimensions given by length and width, standard location and orientation.

Examples

julia> Triangle(;length = 1.0, width = 1.0);

PlantGeomPrimitives.X — Method

X(s)

Returns scaled vector in the direction of the X axis with length s as a Vec object using the same floating point precision as s.

julia>  X(1.0);

julia>  X(1f0) ;

PlantGeomPrimitives.X — Method

X()

Returns an unit vector in the direction of the X axis as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in X(Float32).

julia>  X();

julia>  X(Float32);

PlantGeomPrimitives.Y — Method

Y(s)

Returns scaled vector in the direction of the Y axis with length s as a Vec object using the same floating point precision as s.

julia>  Y(1.0);

julia>  Y(1f0);

PlantGeomPrimitives.Y — Method

Y()

Returns an unit vector in the direction of the Y axis as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in Y(Float32).

julia>  Y();

julia>  Y(Float32);

PlantGeomPrimitives.Z — Method

Z(s)

Returns scaled vector in the direction of the Z axis with length s as a Vec object using the same floating point precision as s.

julia>  Z(1.0);

julia>  Z(1f0);

PlantGeomPrimitives.Z — Method

Z()

Returns an unit vector in the direction of the Z axis as a Vec object. By default, the coordinates will be in double floating precision (Float64) but it is possible to generate a version with lower floating precision as in Z(Float32).

julia>  Z();

julia>  Z(Float32);

PlantGeomPrimitives.add! — Method

add!(scene; mesh, color = nothing, material = nothing)

Manually add a 3D mesh to an existing Scene object (scene) with optional colors and materials

PlantGeomPrimitives.area — Method

area(m::Mesh)

Total surface area of a mesh (as the sum of areas of individual triangles).

PlantGeomPrimitives.areas — Method

areas(m::Mesh)

A vector with the areas of the different triangles that form a mesh.

PlantGeomPrimitives.colors — Method

colors(scene::Scene)

Extract the vector of Colorant objects stored inside a scene (used for rendering)

PlantGeomPrimitives.load_mesh — Method

load_mesh(filename)

Import a mesh from a file given by filename. Supported formats include stl, ply, obj and msh. By default, this will generate a Mesh object that uses double floating-point precision. However, a lower precision can be specified by passing the relevant data type as in load_mesh(filename, Float32).

PlantGeomPrimitives.materials — Method

materials(scene::Scene)

Extract the vector of Material objects stored inside a scene (used for ray tracing)

PlantGeomPrimitives.mesh — Method

mesh(scene::Scene)

Extract the triangular mesh stored inside a scene (used for ray tracing & rendering)

PlantGeomPrimitives.ntriangles — Method

ntriangles(mesh)

Extract the number of triangles in a mesh.

PlantGeomPrimitives.nvertices — Method

nvertices(mesh)

The number of vertices in a mesh.

PlantGeomPrimitives.rotate! — Method

rotate!(m::Mesh; x::Vec, y::Vec, z::Vec)

Rotate a mesh m to a new coordinate system given by x, y and z

PlantGeomPrimitives.rotatex! — Method

rotatex!(m::Mesh, θ)

Rotate a mesh m around the x axis by θ rad.

PlantGeomPrimitives.rotatey! — Method

rotatey!(m::Mesh, θ)

Rotate a mesh m around the y axis by θ rad.

PlantGeomPrimitives.rotatez! — Method

rotatez!(m::Mesh, θ)

Rotate a mesh m around the z axis by θ rad.

PlantGeomPrimitives.save_mesh — Method

save_mesh(mesh; fileformat = STL_BINARY, filename)

Save a mesh into an external file using a variety of formats.

Arguments

mesh: Object of type Mesh.

fileformat: Format to store the mesh. This is a keyword argument.

filename: Name of the file in which to store the mesh.

Details

The fileformat should take one of the following arguments: STL_BINARY, STL_ASCII, PLY_BINARY, PLY_ASCII or OBJ. Note that these names should not be quoted as strings.

Return

This function does not return anything, it is executed for its side effect.

PlantGeomPrimitives.scale! — Method

scale!(m::Mesh, Vec)

Scale a mesh m along the three axes provided by vec

PlantGeomPrimitives.translate! — Method

translate!(m::Mesh, v::Vec)

Translate the mesh m by vector v

Private

Private functions, types or constants from PlantGeomPrimitives. These are not exported, so you need to prefix the function name with PlantGeomPrimitives. to access them. Also bear in mind that these are not part of the public API, so they may change without notice.